requirements.txt
asgiref==3.5.2
autopep8==1.6.0
bitrix24-python3-client==0.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Deprecated==1.2.13
distro==1.7.0
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
netaddr==0.8.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.0
Pygments==2.12.0
PyJWT==2.4.0
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.11
wrapt==1.14.1
yookassa==2.3.2
